% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exercise-model.R
\name{run_exercise_model}
\alias{run_exercise_model}
\title{Run the exercise model}
\usage{
run_exercise_model(
  s = 0.2,
  k = 2.5,
  n = 1,
  i = 1.5,
  j = 1.5,
  a = c(1, 5, 7),
  debug = FALSE
)
}
\arguments{
\item{s}{the scale of the positive impact of activity. Decreasing s makes this impact
slower decaying.}

\item{k}{impact of activity on positive effect}

\item{n}{the dampening of the positive effect of activity.}

\item{i}{numeric. final scaling factor of negative effect}

\item{j}{numeric. final scaling factor of positive effect}

\item{a}{numeric vector. Beginning of activities (all take time 1)}

\item{debug}{show some debugging output.}
}
\value{
a data frame containing six variables. The first column contains the time, and
the final columm contains the final TRI.
}
\description{
Run the ODE model
run_exercise_model() defaults are s=0.2,k=2.5,n=1,i=1.5,j=1.5, a=c(1,5,7)
or choose own values with
run_exercise_model(s=0.3,k=2,n=1.2,i=1,j=1.7, a=c(1,3,5,7,9))
}
\examples{
# Run model with default parameters
x <- run_exercise_model()
plot( TRI ~ time, x, type='l' )
}
